6-8测试优化,修复角色分配用户，登录时用户无角色信息

+++ cloud-provider-rszy8003/src/main/java/com/hjy/cloud/t_outfit/service/impl/TOutfitDeptServiceImpl.java
@@ -116,6 +116,9 @@ public class TOutfitDeptServiceImpl implements TOutfitDeptService {
     @Transactional()
     @Override
     public CommonResult delete(TOutfitDept tOutfitDept) {
+        if("1".equals(tOutfitDept.getPkDeptId())){
+            return new CommonResult().ErrorResult("分公司部门不可删除!",null);
+        }
         int i = this.tOutfitDeptMapper.deleteById(tOutfitDept);
         String msg = "部门数据删除成功!";
         if (i > 0) {
